### 2026-05-31 #108 @DerKevinRiehl
[RE]  A Cyber-Physical System for Freeway Ramp Meter Signal Control Using Deep Reinforcement Learning in a Connected Environment
**Original article:** Hou, Y., Zhang, X., Graf, P., Tripp, C., & Biagioni, D. (2021). A Cyber-Physical System for Freeway Ramp Meter Signal Control Using Deep Reinforcement Learning in a Connected Environment. In 2021 IEEE Intelligent Transportation Systems Conference (ITSC), pp. 1--8." doi: 10.1109/ITSC48978.2021.9564699.

**Scientific domain:** Reinforcement Learning Based Control
**Suggested editor:** @rougier
